[fix] Split EProjectManager.cs to single files
